yaml_to_lean_translation/requirements_translation.txt
anthropic>=0.18.0
pyyaml>=6.0
pathlib2>=2.3.7
